+++ feature/notice/build.gradle.kts
@@ -26,6 +26,7 @@ dependencies {
     implementation(project(":core:designsystem"))
     implementation(project(":core:domain"))
     implementation(project(":core:model"))
+    implementation(project(":core:common"))
 
     implementation(libs.bundles.androidx)
     implementation(libs.material)
